(feat) create the 'AbstractSwipeSection' to prevent code duplication on the swipe pages
