+++ src/app/core/shared/swipe.section.ts
@@ -1,0 +1,12 @@
+export declare interface SwipeSection {
+
+    onSwipe(event: any): void;
+
+    disablePreviousNavigation(): boolean;
+
+    disableNextNavigation(): boolean;
+
+    onClickPrevious(): void;
+
+    onClickNext(): void;
+}
